feat(ARC-3815): refactor timeline object rendering and styling
Separates the media display and metadata/CTA logic for timeline objects into dedicated components (`BlockTimelineObject` and `BlockTimelineObjectMeta`). This improves modularity and maintainability of the `BlockTimeline`.

Updates styling for responsive year column widths and refines the object metadata layout for a clearer visual presentation.

https://meemoo.atlassian.net/browse/ARC-3815
